+++ CHANGELOG.md
@@ -5,9 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Update `Alert` component with new design styleguide V2.
+- Feature: Add `disabled` prop on `TextInputWithButton`.
+- Feature: Update all `Input` components with new design styleguide V2.
+- Feature: Update `k-buttonRounded` helper to manage button radius.
 - Feature: Add `k-buttonPadding` helper to manage button paddings.
 - Feature: Update all `Button` components with new design styleguide V2.
-- Fix: Update marge on `FormComposer` component.
+- Fix: Update ruby version.
 - Feature: Remove `boron` buttons colors.
 - Feature: Update colors.
 - Feature: Add new `error` and `valid` colors.